[Encoder] Refactore content of encoder.rs.
Renames encoder.rs into encoding.rs and splits its content into
multiple submodules to highlight the difference between the actual
`Encoder` implementation and the highlevel `Encodable` trait.

To avoid any kind of breaking changes this keeps the exported module
name `encoder` even if it internally renames it into `encoding`.
