+++ src/encoding/encoder.rs
@@ -1,136 +1,10 @@
-//! An encoder for bencode. Guarantees that the output string is valid bencode
-//!
-//! # Encoding a structure
-//!
-//! The easiest way to encode a structure is to implement [`Encodable`] for it. For most structures,
-//! this should be very simple:
-//!
-//! ```
-//! # use bendy::encoder::{Encodable, SingleItemEncoder};
-//! # use bendy::Error;
-//!
-//! struct Message {
-//!     foo: i32,
-//!     bar: String,
-//! }
-//!
-//! impl Encodable for Message {
-//!     // Atoms have depth one. The struct wrapper adds one level to that
-//!     const MAX_DEPTH: usize = 1;
-//!
-//!     fn encode(&self, encoder: SingleItemEncoder) -> Result<(), Error> {
-//!         encoder.emit_dict(|mut e| {
-//!             // Use e to emit the values
-//!             e.emit_pair(b"bar", &self.bar)?;
-//!             e.emit_pair(b"foo", &self.foo)
-//!         })
-//!     }
-//! }
-//! ```
-//!
-//! Then, messages can be serialized using [`Encodable::to_bytes`]:
-//!
-//! ```
-//! # use bendy::encoder::{Encodable, SingleItemEncoder};
-//! # use bendy::Error;
-//! #
-//! # struct Message {
-//! #    foo: i32,
-//! #    bar: String,
-//! # }
-//! #
-//! # impl Encodable for Message {
-//! #     // Atoms have depth zero. The struct wrapper adds one level to that
-//! #     const MAX_DEPTH: usize = 1;
-//! #
-//! #     fn encode(&self, encoder: SingleItemEncoder) -> Result<(), Error> {
-//! #         encoder.emit_dict(|mut e| {
-//! #             // Use e to emit the values. They must be in sorted order here.
-//! #             // If sorting the dict first is annoying, you can also use
-//! #             // encoder.emit_and_sort_dict
-//! #             e.emit_pair(b"bar", &self.bar)?;
-//! #             e.emit_pair(b"foo", &self.foo)
-//! #         })
-//! #     }
-//! # }
-//! # let result: Result<Vec<u8>, Error> =
-//! Message{
-//!     foo: 1,
-//!     bar: "quux".to_string(),
-//! }.to_bytes()
-//! # ;
-//! # assert!(result.is_ok());
-//! ```
-//!
-//! Most primitive types already implement [`Encodable`].
-//!
-//! # Nesting depth limits
-//!
-//! To allow this to be used on limited platforms, all implementations of [`Encodable`] include a
-//! maximum nesting depth. Atoms (integers and byte strings) are considered to have depth 0. An
-//! object (a list or dict) containing only atoms has depth 1, and in general, an object has a depth
-//! equal to the depth of its deepest member plus one. In some cases, an object doesn't have a
-//! statically known depth. For example, ASTs may be arbitrarily nested. Such objects should
-//! have their depth set to 0, and callers should construct the Encoder manually, adding an
-//! appropriate buffer for the depth:
-//!
-//! ```
-//! # use bendy::encoder::{Encodable, Encoder};
-//! # use bendy::Error;
-//! #
-//! # type ObjectType = u32;
-//! # static object: u32 = 0;
-//! #
-//! # fn main() -> Result<(), Error> {
-//! let mut encoder = Encoder::new()
-//!     .with_max_depth(ObjectType::MAX_DEPTH + 10);
-//! encoder.emit(object)?;
-//! encoder.get_output()
-//! #   .map(|_| ()) // ignore a success return value
-//! # }
-//! ```
-//!
-//! # Error handling
-//!
-//! Once an error occurs during encoding, all future calls to the same encoding stream will fail
-//! early with the same error. It is not defined whether any callback or implementation of
-//! [`Encodable::encode`] is called before returning an error; such callbacks should respond to
-//! failure by bailing out as quickly as possible.
-//!
-//! Not all values in [`Error`] can be caused by an encoding operation. Specifically, you only need
-//! to worry about [`UnsortedKeys`] and [`NestingTooDeep`].
-//!
-//! [`UnsortedKeys`]: self::Error#UnsortedKeys
-//! [`NestingTooDeep`]: self::Error#NestingTooDeep
-
-use std::{
-    collections::{BTreeMap, HashMap, LinkedList, VecDeque},
-    hash::Hash,
-    io::{self, Write},
-};
+use std::{collections::BTreeMap, io::Write};
 
 use crate::{
+    encoding::{Encodable, PrintableInteger},
     state_tracker::StateTracker,
     token::{Error, Token},
 };
-
-/// A value that can be formatted as a decimal integer
-pub trait PrintableInteger {
-    /// Write the value as a decimal integer
-    fn write_to<W: Write>(self, w: W) -> io::Result<()>;
-}
-
-macro_rules! impl_integer {
-    ($($type:ty)*) => {$(
-        impl PrintableInteger for $type {
-            fn write_to<W: Write>(self, mut w: W) -> io::Result<()> {
-                write!(w, "{}", self)
-            }
-        }
-    )*}
-}
-
-impl_integer!(u8 u16 u32 u64 usize i8 i16 i32 i64 isize);
 
 /// The actual encoder. Unlike the decoder, this is not zero-copy, as that would
 /// result in a horrible interface
@@ -518,197 +392,10 @@ impl UnsortedDictEncoder {
     }
 }
 
-/// An object that can be encoded into a single bencode object
-pub trait Encodable {
-    /// The maximum depth that this object could encode to. Leaves do not consume a level, so an
-    /// `i1e` has depth 0 and `li1ee` has depth 1.
-    const MAX_DEPTH: usize;
-
-    /// Encode this object into the bencode stream
-    fn encode(&self, encoder: SingleItemEncoder) -> Result<(), Error>;
-
-    /// Encode this object to a byte string
-    fn to_bytes(&self) -> Result<Vec<u8>, Error> {
-        let mut encoder = Encoder::new().with_max_depth(Self::MAX_DEPTH);
-        encoder.emit_with(|e| self.encode(e))?;
-        encoder.get_output()
-    }
-}
-
-// Forwarding impls
-impl<'a, E: 'a + Encodable + Sized> Encodable for &'a E {
-    const MAX_DEPTH: usize = E::MAX_DEPTH;
-
-    fn encode(&self, encoder: SingleItemEncoder) -> Result<(), Error> {
-        E::encode(self, encoder)
-    }
-}
-
-impl<E: Encodable> Encodable for Box<E> {
-    const MAX_DEPTH: usize = E::MAX_DEPTH;
-
-    fn encode(&self, encoder: SingleItemEncoder) -> Result<(), Error> {
-        E::encode(&*self, encoder)
-    }
-}
-
-impl<E: Encodable> Encodable for ::std::rc::Rc<E> {
-    const MAX_DEPTH: usize = E::MAX_DEPTH;
-
-    fn encode(&self, encoder: SingleItemEncoder) -> Result<(), Error> {
-        E::encode(&*self, encoder)
-    }
-}
-
-impl<E: Encodable> Encodable for ::std::sync::Arc<E> {
-    const MAX_DEPTH: usize = E::MAX_DEPTH;
-
-    fn encode(&self, encoder: SingleItemEncoder) -> Result<(), Error> {
-        E::encode(&*self, encoder)
-    }
-}
-
-// Base type impls
-impl<'a> Encodable for &'a str {
-    const MAX_DEPTH: usize = 0;
-
-    fn encode(&self, encoder: SingleItemEncoder) -> Result<(), Error> {
-        encoder.emit_str(self)
-    }
-}
-
-impl Encodable for String {
-    const MAX_DEPTH: usize = 0;
-
-    fn encode(&self, encoder: SingleItemEncoder) -> Result<(), Error> {
-        encoder.emit_str(self)
-    }
-}
-
-macro_rules! impl_encodable_integer {
-    ($($type:ty)*) => {$(
-        impl Encodable for $type {
-            const MAX_DEPTH: usize = 1;
-
-            fn encode(&self, encoder: SingleItemEncoder) -> Result<(), Error> {
-                encoder.emit_int(*self)
-            }
-        }
-    )*}
-}
-
-impl_encodable_integer!(u8 u16 u32 u64 usize i8 i16 i32 i64 isize);
-
-macro_rules! impl_encodable_iterable {
-    ($($type:ident)*) => {$(
-        impl <ContentT> Encodable for $type<ContentT>
-        where
-            ContentT: Encodable
-        {
-            const MAX_DEPTH: usize = ContentT::MAX_DEPTH + 1;
-
-            fn encode(&self, encoder: SingleItemEncoder) -> Result<(), Error> {
-                encoder.emit_list(|e| {
-                    for item in self {
-                        e.emit(item)?;
-                    }
-                    Ok(())
-                })
-            }
-        }
-    )*}
-}
-
-impl_encodable_iterable!(Vec VecDeque LinkedList);
-
-impl<'a, ContentT> Encodable for &'a [ContentT]
-where
-    ContentT: Encodable,
-{
-    const MAX_DEPTH: usize = ContentT::MAX_DEPTH + 1;
-
-    fn encode(&self, encoder: SingleItemEncoder) -> Result<(), Error> {
-        encoder.emit_list(|e| {
-            for item in *self {
-                e.emit(item)?;
-            }
-            Ok(())
-        })
-    }
-}
-
-impl<K: AsRef<[u8]>, V: Encodable> Encodable for BTreeMap<K, V> {
-    const MAX_DEPTH: usize = V::MAX_DEPTH + 1;
-
-    fn encode(&self, encoder: SingleItemEncoder) -> Result<(), Error> {
-        encoder.emit_dict(|mut e| {
-            for (k, v) in self {
-                e.emit_pair(k.as_ref(), v)?;
-            }
-            Ok(())
-        })
-    }
-}
-
-impl<K, V, S> Encodable for HashMap<K, V, S>
-where
-    K: AsRef<[u8]> + Eq + Hash,
-    V: Encodable,
-    S: ::std::hash::BuildHasher,
-{
-    const MAX_DEPTH: usize = V::MAX_DEPTH + 1;
-
-    fn encode(&self, encoder: SingleItemEncoder) -> Result<(), Error> {
-        encoder.emit_dict(|mut e| {
-            let mut pairs = self
-                .iter()
-                .map(|(k, v)| (k.as_ref(), v))
-                .collect::<Vec<_>>();
-            pairs.sort_by_key(|&(k, _)| k);
-            for (k, v) in pairs {
-                e.emit_pair(k, v)?;
-            }
-            Ok(())
-        })
-    }
-}
-
-/// Wrapper to allow `Vec<u8>` encoding as bencode string element.
-#[derive(Clone, Copy, Debug, Default, Hash, Eq, PartialEq, PartialOrd, Ord)]
-pub struct AsString<I>(pub I);
-
-impl<I> Encodable for AsString<I>
-where
-    I: AsRef<[u8]>,
-{
-    const MAX_DEPTH: usize = 1;
-
-    fn encode(&self, encoder: SingleItemEncoder) -> Result<(), Error> {
-        encoder.emit_bytes(self.0.as_ref())
-    }
-}
-
-impl<I> AsRef<[u8]> for AsString<I>
-where
-    I: AsRef<[u8]>,
-{
-    fn as_ref(&self) -> &'_ [u8] {
-        self.0.as_ref()
-    }
-}
-
-impl<'a, I> From<&'a [u8]> for AsString<I>
-where
-    I: From<&'a [u8]>,
-{
-    fn from(content: &'a [u8]) -> Self {
-        AsString(I::from(content))
-    }
-}
-
 #[cfg(test)]
 mod test {
     use super::*;
+
     #[test]
     pub fn simple_encoding_works() {
         let mut encoder = Encoder::new();
@@ -731,45 +418,9 @@ mod test {
         );
     }
 
-    struct Foo {
-        bar: u32,
-        baz: Vec<String>,
-        qux: Vec<u8>,
-    }
-
-    impl Encodable for Foo {
-        const MAX_DEPTH: usize = 2;
-
-        fn encode(&self, encoder: SingleItemEncoder) -> Result<(), Error> {
-            encoder.emit_dict(|mut e| {
-                e.emit_pair(b"bar", &self.bar)?;
-                e.emit_pair(b"baz", &self.baz)?;
-                e.emit_pair(b"qux", AsString(&self.qux))?;
-                Ok(())
-            })
-        }
-    }
-
-    #[test]
-    fn simple_encodable_works() {
-        let mut encoder = Encoder::new();
-        encoder
-            .emit(Foo {
-                bar: 5,
-                baz: vec!["foo".to_owned(), "bar".to_owned()],
-                qux: b"qux".to_vec(),
-            })
-            .unwrap();
-        assert_eq!(
-            &encoder.get_output().unwrap()[..],
-            &b"d3:bari5e3:bazl3:foo3:bare3:qux3:quxe"[..]
-        );
-    }
-
     #[test]
     fn emit_cb_must_emit() {
         let mut encoder = Encoder::new();
-
         assert!(encoder.emit_with(|_| Ok(())).is_err());
     }
 }
